bug: fixes updateissueSatus resolver user authorization
