--- src/resolvers/Mutation/changeIssueStatus.ts
@@ -1,120 +1,0 @@
-import _ from 'lodash';
-import { MutationResolvers } from '../../types/generatedGraphQLTypes';
-import { IssueStatus, IssueType, UserTeam } from '@prisma/client';
-import { notFoundError } from '../../libraries/errors/notFoundError';
-import { conflictError } from '../../libraries/errors/conflictError';
-import { UNAUTHORIZED_USER, ISSUE_NOT_FOUND, PROJECT_NOT_FOUND } from '../../globals';
-
-export const updateIssueStatus: MutationResolvers['updateIssueStatus'] = async (
-  _,
-  args,
-  context
-) => {
-
-  try {
-
-    if (!context.userId) {
-      throw new conflictError(UNAUTHORIZED_USER.message, UNAUTHORIZED_USER.code);
-    }
-
-    const project = await context.client.project.findUnique({
-      where: {
-        id: args.projectId,
-      },
-      select: {
-        id: true,
-        creatorId: true,
-        teams: {
-          select: {
-            team: {
-              select: {
-                users: {
-                  where: {
-                    id: context.userId,
-                  },
-                },
-              },
-            },
-          },
-        },
-      },
-    });
-
-    if (!project) {
-      throw new notFoundError(PROJECT_NOT_FOUND.MESSAGE, PROJECT_NOT_FOUND.CODE);
-    }
-
-    const isMember = project.teams.some((team: any) =>
-      team.team.users.some((user: UserTeam) => user.id === context.userId)
-    );
-
-    if (!isMember) {
-      throw new conflictError(UNAUTHORIZED_USER.message, UNAUTHORIZED_USER.code);
-    }
-
-    if (args.status === IssueStatus.DONE) {
-      const issue = await context.client.issue.findUnique({
-        where: {
-          id: args.issueId,
-        },
-        select: {
-          id: true,
-          type: true,
-          childrens: {
-            select: {
-              id: true,
-              status: true,
-              childrens: {
-                select: {
-                  id: true,
-                  status: true,
-                }
-              }
-            }
-          }
-        }
-      });
-
-      if (!issue) {
-        throw new notFoundError(ISSUE_NOT_FOUND.MESSAGE, ISSUE_NOT_FOUND.CODE);
-      }
-
-      // Recursively check if all children are DONE
-      function allChildrenDone(children: any[]): boolean {
-        return children.every(child =>
-          child.status === IssueStatus.DONE &&
-          (!child.childrens || allChildrenDone(child.childrens))
-        );
-      }
-
-      if (issue.childrens && issue.childrens.length > 0 && !allChildrenDone(issue.childrens)) {
-        throw new conflictError(
-          "Children Issues are not DONE yet",
-          "issue.childNotDone"
-        );
-      }
-    }
-
-    try {
-      await context.client.$transaction(async (t) => {
-        await t.issue.update({
-          where: {
-            id: args.issueId,
-          },
-          data: {
-            status: args.status,
-          },
-        });
-      });
-    } catch (e) {
-      throw new conflictError('Unable to Update Issue', 'issue.updateFailed');
-    }
-  } catch (e) {
-    throw new Error('Unable to Update Issue');
-  }
-
-  return {
-    message: 'Issue Updated',
-    success: true,
-  };
-};
